+++ test/manual/src/custom-square.js
@@ -1,38 +1,38 @@
 /* eslint-env browser */
 /* eslint-disable */
 
-// import customSquareStyles from './custom-square.lazy.css';
+import customSquareStyles from "./custom-square.custom.css";
 
 class CustomSquare extends HTMLElement {
   // Specify observed attributes so that
   // attributeChangedCallback will work
   static get observedAttributes() {
-    return ['l'];
+    return ["w", "h"];
   }
 
   constructor() {
     // Always call super first in constructor
     super();
 
-    this.attachShadow({ mode: 'open' });
-  }
-  connectedCallback() {
-    const div = document.createElement('div');
+    this.attachShadow({ mode: "open" });
+    const divElement = document.createElement("div");
 
-    this.shadowRoot.appendChild(div);
+    divElement.textContent = "Text content.";
+
+    this.shadowRoot.appendChild(divElement);
+
+    customSquareStyles.use({ target: this.shadowRoot });
 
     const bgPurple = new CSSStyleSheet();
+    const width = this.getAttribute("w");
+    const height = this.getAttribute("h");
 
-    bgPurple.replace(`div { 
-    background: purple; 
-    width: ${this.getAttribute('l')}px;
-    height: ${this.getAttribute('l')}px;
-}`);
+    bgPurple.replace(`div { width: ${width}px; height: ${height}px; }`);
 
     this.shadowRoot.adoptedStyleSheets = [bgPurple];
   }
 }
 
-customElements.define('custom-square', CustomSquare);
+customElements.define("custom-square", CustomSquare);
 
 export default CustomSquare;